vbswap: Introduce VBSWAP_HELPER
Based on celtare21/kernel_google_coral@8e14254:

This is a simplified driver designed to assist in the configuration
and initialization of VBSWAP while avoiding ramdisk dependencies.

This is achieved by temporarily disabling selinux, similiar in
methodology to execprog, but less invasive, and with a much narrower
scope. In doing so, we are able to call userspace functions from
kernel and subsequently configure default VBSWAP disksize, as well
as perform the mkswap/swapon operations necessary to initialize the
device.

Co-authored-by: celtare21 <celtare21@gmail.com>
[@0ctobot: Adapt and refactor userland_worker as a dedicated swap helper]
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>
Signed-off-by: Zlatan Radovanovic <zlatan.radovanovic@fet.ba>